Fix for critical GKE/GCP Terraform Bugs
This fixes several critical issues, and also cleans up other aspect
along the way. Primarily that the firewall rule would never allow
traffic to the cluster, since they were erroneously configured and on a
completely different network to the cluster.

Bug Fixes:
1. Remove the creation of a separate network for each cluster
1. Make sure the firewall rules don't filter on tags, but target on
   network tags.
1. Be able to pass in a `network` argument (defaulting to "default"),
such that if you want to run a cluster on a different network it is easy
to configure - this means that the firewall rule is always on the same
network as the cluster.
1. Controller deployment was set to "Always" image pull policy on both
the controller and the sidecar. This should never be configured this way
for production systems, and is therefore an unsafe default.

Cleanup:
1. Remove duplicate helm.tf and cluster.tf in root ./install/terraform
path, as they were diverging, and served no value. Both ./gke and
./gke-local installation scripts have also been updated, as well as the
documentation.
1. Needed to upgrade Terraform in the build shell, since all modules
used a later version.
1. Cleaned up deprecation comments
1. Removed the broken symlinks in ./build

I would like to also make it easier to have only one GCP firewall rule
for n number of clusters, but we can tackle that in a later PR.

Closes #1372